+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Offer functionality through library in addition to command line.
+  - The CLI functionality requires the on-by-default "cli" feature.
+
 ### Changed
 - Set LTO for release builds by default